Simplify full/partial completion check.
Since we already depend on the knowledge that we do not need UTF-8 in this
place as the comment says, we may as well depend on the fact that the
regexp is anchored to the beginning of line.  Cutting by byte offset won't
work with Term::ReadLine::Gnu 1.29+ anyway as noted in the previous
comment.